backend/requirements.txt
fastapi
uvicorn[standard]
catboost
numpy
spacy
sqlalchemy
alembic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings
email-validator
